feat: refactor MapCity component into modular files
- Deleted the old MapCity.vue file and created a new modular structure with separate files for JavaScript (map-city.js), CSS (map-city.css), and the Vue template.
- Implemented a responsive design for the map display and category selection.
- Enhanced user experience with loading indicators and improved map functionalities using Gmap components.
